Add ability to specify which field holds build info.
Made the task multi-task, primarily to support testing.

Add a buildField option, to specify which field the build
info is added to.  This allows avoidance of the default 'build'
field which is a future reserved field in npm package.json.

Update tests to make a copy of the actual data each time and
modify it, rather than changing the versioned controlled copies.